stats_service/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
clickhouse-driver==0.2.5
kafka-python==2.1.5
python-dotenv==0.19.2
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0